--- src/components/common/guidePage/guidePage.scss
@@ -1,22 +1,0 @@
-.guide-body {
-    height: calc(100vh - 454px);
-}
-
-.guide-container {
-    .guide-cards__wrap {
-        display: grid;
-        grid-template-columns: 300px 300px;
-        grid-gap: 20px;
-        position: absolute;
-        top: -80px;
-    }
-
-    .guide-card__img {
-        width: 298px;
-        height: 210px;
-    }
-
-    .guide-card:hover{
-      box-shadow: 0 8px 12px 0 rgba(30, 35, 96, 0.1), 0 1px 4px 0 rgba(0, 0, 0, 0.1);
-    }
-}
