+++ src/css/colorPalette.css
@@ -48,4 +48,5 @@
     --orange: rgba(252, 173, 38, 1);
     --window-bg: #f2f4f7;
     --white: #fff;
+    --border: #1e23601a
 }
